Rename services namespace for consistency
