+++ Epsilon.Canvas.Abstractions/Service/IAssignmentService.cs
@@ -1,6 +1,6 @@
 ﻿using Epsilon.Canvas.Abstractions.Model;
 
-namespace Epsilon.Canvas.Abstractions.Services;
+namespace Epsilon.Canvas.Abstractions.Service;
 
 public interface IAssignmentService
 {
