+++ Epsilon.Canvas.Abstractions/Service/IOutcomeService.cs
@@ -1,6 +1,6 @@
 ﻿using Epsilon.Canvas.Abstractions.Model;
 
-namespace Epsilon.Canvas.Abstractions.Services;
+namespace Epsilon.Canvas.Abstractions.Service;
 
 public interface IOutcomeService
 {
