+++ Epsilon.Canvas/CanvasServiceCollectionExtensions.cs
@@ -1,6 +1,6 @@
 using System.Net.Http.Headers;
 using Epsilon.Canvas.Abstractions;
-using Epsilon.Canvas.Abstractions.Services;
+using Epsilon.Canvas.Abstractions.Service;
 using Epsilon.Canvas.Service;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
