+++ Epsilon.Canvas/Service/OutcomeService.cs
@@ -1,5 +1,5 @@
 ﻿using Epsilon.Canvas.Abstractions.Model;
-using Epsilon.Canvas.Abstractions.Services;
+using Epsilon.Canvas.Abstractions.Service;
 using Epsilon.Canvas.Response;
 using Epsilon.Http.Abstractions;
 using Epsilon.Http.Abstractions.Json;
