+++ Epsilon.Canvas/Service/SubmissionService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text;
 using Epsilon.Canvas.Abstractions.Model;
-using Epsilon.Canvas.Abstractions.Services;
+using Epsilon.Canvas.Abstractions.Service;
 using Epsilon.Http.Abstractions;
 using Epsilon.Http.Abstractions.Json;
